UPSTREAM: 75126: available controller: do not stomp transport
We should not stomp on the discovery transport with the proxy
transport as the proxy transport is always unauthenticated.  We need
to make sure these calls are always authenticated as system:masters
to prevent any forbidden errors.  Thus we only use the proxy
transport to set the dial context.

This matches the logic upstream (while maintaining our dynamic cert
reloading logic):

https://github.com/kubernetes/kubernetes/blob/d873167e8f816f5f83846bc4587c3047452293bd/staging/src/k8s.io/kube-aggregator/pkg/controllers/status/available_controller.go#L114-L116

Signed-off-by: Monis Khan <mkhan@redhat.com>

Origin-commit: 3876e542a6bb37f9db0936af986934961bdb7172

Kubernetes-commit: 96ed8ca93d064bfa2b1dce77aee0214f2eaddf7f